astacus: move memory object storage to tests
MemoryObjectStorage is considered for testing purpose
